requirements.txt
certifi==2020.6.20
click==7.1.2
cloudinary==1.22.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dotenv==0.14.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.10
Werkzeug==1.0.1
